change type of mutationprobability from double to unsigned int, start doxygen comments

+++ src/GameMain/Gamemaster.cc
@@ -11,7 +11,7 @@ using namespace std;
 
 Gamemaster::Gamemaster(){};
 
-Gamemaster::Gamemaster(unsigned int k, unsigned int length, unsigned int number_of_sequences, double probability){
+Gamemaster::Gamemaster(unsigned int k, unsigned int length, unsigned int number_of_sequences, unsigned int probability){
     simulate(number_of_sequences, length, probability);     
     GameGraph.readFastaFiles("sequences.fa", k);
     remove("sequences.fa");
